Move fuzzy search function to core common utilities
- Deleted the previous implementation of the fuzzy search function from StringExtensions.kt in the app module.
- Added the fuzzy search function to StringExtensions.kt in the core common module, ensuring consistent usage across the application.
- This change improves code maintainability and centralizes the search logic for better performance.
